Speech Processing Project
Fine tuning of WAV2VEC2 model on Urdu speech of 5 participants (dataset not uploaded). 

The model was already trained on English speech datasets. We attempted to fine tune the dataset on high quality Urdu speech data, to test the feasibility of expanding the usage of the model on low-resource languages given high quality but limited data.

The dataset focused on representing all possible combinations of Urdu tri-phoneme combinations. A tri-phoneme is any combination of any 3 phonemes within a language. This definition considers silence as a phoneme.